+++ config/initializers/onesignal.rb
@@ -1,0 +1,1 @@
+OneSignal::OneSignal.api_key = ENV["ONESIGNAL_API_KEY"]